pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gamma-skill-studio"
version = "0.1.0"
description = "A web studio for authoring humanoid motion libraries."
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [{ name = "Gamma Lab", email = "hello@gammalab.ae" }]
requires-python = ">=3.10"
keywords = ["robotics", "humanoid", "motion", "urdf", "motion-library"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "fastapi>=0.110,<1.0",
    "uvicorn[standard]>=0.27,<1.0",
    "pydantic>=2.5,<3.0",
    "pyyaml>=6.0,<7.0",
    "numpy>=1.24,<3.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4,<9.0", "httpx>=0.25,<1.0", "ruff>=0.3,<1.0"]

[project.urls]
Homepage = "https://github.com/Gamma-Humanoids/gamma-skill-studio"
Issues = "https://github.com/Gamma-Humanoids/gamma-skill-studio/issues"
"Gamma Lab" = "https://gammalab.ae/Home"

[tool.setuptools]
packages = []

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
